.history/requirements_20251223235729.txt
python>=3.8

numpy
pandas
matplotlib
pillow
scipy
torch
torchvision
